pyproject.toml
[project]
name = "lightrag-backend"
version = "0.1.0"
description = "Text-only RAG backend using LightRAG and RAGAnything"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.28.0",
    "lightrag-hku[all]>=1.4.9.11",
    "pymupdf4llm>=0.3.4",
    "pymupdf>=1.24.0",
    "pymilvus>=2.3.6",
    "pydantic-settings>=2.2.1",
    "structlog>=24.1.0",
]
